rename package name to keep the same with dir name
