+++ keadm/cmd/keadm/app/cmd/cloud/gettoken.go
@@ -1,4 +1,4 @@
-package cmd
+package cloud
 
 import (
 	"context"
